refact: zustand 좌석, 리뷰 스토어 분리
